requirements.txt
boto3>=1.40.52
pyyaml>=6.0.3
texttable>=1.7.0
jinja2>=3.1.6
cfn-lint>=1.40.2
moto[all]>=5.1.14 # Unittests
coverage>=7.11.0
